+++ src/reader/__init__.py
@@ -37,7 +37,7 @@ Print the titles of the unread entries::
 
 """
 
-__version__ = '0.11'
+__version__ = '0.12.dev0'
 
 
 from .core.reader import Reader