+++ CHANGES.rst
@@ -10,7 +10,7 @@ Version 0.10
 
 Unreleased
 
-* Unify plugin loading code and error handling. (`#112`_)
+* Unify plugin loading and error handling code. (`#112`_)
 * Minor improvements to CLI error reporting.
 
 .. _#112: https://github.com/lemon24/reader/issues/112