+++ CHANGES.rst
@@ -5,6 +5,12 @@ reader changelog
 .. module:: reader
 
 
+Version 0.12
+------------
+
+Unreleased
+
+
 Version 0.11
 ------------